requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
flake8==7.1.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
smmap==5.0.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.13.1
